requirements.txt
scipy==1.7.1
sounddevice==0.4.2
pydub==0.25.1
librosa==0.8.1
SoundFile==0.10.3.post1
streamlit==0.89.0
numpy==1.20.0
Pillow==8.3.2
